some more epurations

+++ source_files/server.c
@@ -6,7 +6,7 @@
 /*   By: dsylvain <dsylvain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 08:45:49 by dsylvain          #+#    #+#             */
-/*   Updated: 2023/12/28 06:44:42 by dsylvain         ###   ########.fr       */
+/*   Updated: 2023/12/28 07:55:57 by dsylvain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,6 +81,8 @@ int	listening_loop(char **input_string)
 		g_server_binary[0] = -1;
 		input_string_len = 0;
 		wait_signal_server();
+		if (g_server_binary[0] == -2)
+			return (0);
 		kill(g_server_binary[1], SIGUSR2);
 		input_string_len = get_string_length_transmission();
 		if (!create_input_string(input_string, input_string_len))
